+++ .github/CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ### New features since last release
 
-* The Jet interpreter for XIR programs now supports state vector outputs via `probability` statements.
+* The Jet interpreter for XIR programs now supports probability distribution outputs via `probabilities` statements.
+
+* The Jet interpreter for XIR scripts now handles gate definitions. [(#37)](https://github.com/XanaduAI/jet/pull/37)
+
+* Full and sliced contractions can now be run with `TaskBasedContractor`on the GPU using the `CudaTensor` class. (#29)](https://github.com/XanaduAI/jet/pull/29)
 
 * CudaTensor now supports `Transpose` calls. [(#42)](https://github.com/XanaduAI/jet/pull/42)